requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-huggingface==0.0.3
langgraph==0.2.28
fastapi==0.110.3
uvicorn==0.29.0
streamlit==1.33.0
sentence-transformers==2.7.0
faiss-cpu==1.8.0
python-dotenv==1.0.0
pydantic==2.7.1
huggingface-hub==0.23.4
requests==2.31.0
python-multipart==0.0.9